Lightcone reading/writing now works
