+++ src/py21cmfast/__init__.py
@@ -9,15 +9,16 @@ from . import cache_tools
 from ._cfg import config
 from ._logging import configure_logging
 from .cache_tools import query_cache
+from .outputs import Coeval
+from .outputs import InitialConditions
+from .outputs import IonizedBox
+from .outputs import LightCone
+from .outputs import PerturbedField
+from .outputs import TsBox
 from .wrapper import AstroParams
 from .wrapper import BrightnessTemp
 from .wrapper import CosmoParams
 from .wrapper import FlagOptions
-from .wrapper import InitialConditions
-from .wrapper import IonizedBox
-from .wrapper import LightCone
-from .wrapper import PerturbedField
-from .wrapper import TsBox
 from .wrapper import UserParams
 from .wrapper import brightness_temperature
 from .wrapper import compute_luminosity_function
