+++ src/py21cmfast/outputs.py
@@ -3,7 +3,7 @@ Output class objects.
 
 The classes provided by this module exist to simplify access to large datasets created within C.
 Fundamentally, ownership of the data belongs to these classes, and the C functions merely accesses
-this and fills it. The various boxes and quantities associated with each output are available as
+this and fills it. The various boxes and lightcones associated with each output are available as
 instance attributes. Along with the output data, each output object contains the various input
 parameter objects necessary to define it.
 
@@ -11,10 +11,20 @@ parameter objects necessary to define it.
              as output objects from the various functions contained here. Only the data
              within the objects should be accessed.
 """
+import os
+import warnings
+from hashlib import md5
+
+import h5py
 import numpy as np
+from astropy import units
+from astropy.cosmology import z_at_value
 from cached_property import cached_property
 
+from . import __version__
+from ._cfg import config
 from ._utils import OutputStruct as _BaseOutputStruct
+from ._utils import _check_compatible_inputs
 from .c_21cmfast import ffi
 from .inputs import AstroParams
 from .inputs import CosmoParams
@@ -351,3 +361,345 @@ class BrightnessTemp(_AllParamsBox):
             )
         else:
             return np.mean(self.brightness_temp)
+
+
+class Coeval:
+    """A full coeval box with all associated data."""
+
+    def __init__(
+        self,
+        redshift: float,
+        init_box: InitialConditions,
+        perturb: PerturbedField,
+        ib: IonizedBox,
+        bt: BrightnessTemp,
+        st: [TsBox, None] = None,
+        photon_nonconservation_data=None,
+    ):
+        _check_compatible_inputs(init_box, perturb, ib, bt, st, ignore=[])
+
+        self.redshift = redshift
+        self.init_struct = init_box
+        self.perturb_struct = perturb
+        self.ionization_struct = ib
+        self.brightness_temp_struct = bt
+        self.spin_temp_struct = st
+        self.photon_nonconservation_data = photon_nonconservation_data
+
+        # Expose all the fields of the structs to the surface of the Coeval object
+        for box in [init_box, perturb, ib, bt]:
+            for field in box.fieldnames:
+                setattr(self, field, getattr(box, field))
+
+    @property
+    def user_params(self):
+        """User params shared by all datasets."""
+        return self.brightness_temp_struct.user_params
+
+    @property
+    def cosmo_params(self):
+        """Cosmo params shared by all datasets."""
+        return self.brightness_temp_struct.cosmo_params
+
+    @property
+    def flag_options(self):
+        """Flag Options shared by all datasets."""
+        return self.brightness_temp_struct.flag_options
+
+    @property
+    def astro_params(self):
+        """Astro params shared by all datasets."""
+        return self.brightness_temp_struct.astro_params
+
+    @property
+    def random_seed(self):
+        """Random seed shared by all datasets."""
+        return self.brightness_temp_struct.random_seed
+
+
+class LightCone:
+    """A full Lightcone with all associate evolved data."""
+
+    def __init__(
+        self,
+        redshift,
+        user_params,
+        cosmo_params,
+        astro_params,
+        flag_options,
+        lightcones,
+        node_redshifts=None,
+        global_quantities=None,
+        photon_nonconservation_data=None,
+        _globals=None,
+    ):
+        self.redshift = redshift
+        self.user_params = user_params
+        self.cosmo_params = cosmo_params
+        self.astro_params = astro_params
+        self.flag_options = flag_options
+        self.node_redshifts = node_redshifts
+
+        # A *copy* of the current global parameters.
+        self.global_params = _globals or dict(global_params.items())
+
+        if global_quantities:
+            for name, data in global_quantities.items():
+                if name.endswith("_box"):
+                    # Remove the _box because it looks dumb.
+                    setattr(self, "global_" + name[:-4], data)
+                else:
+                    setattr(self, "global_" + name, data)
+
+        self.photon_nonconservation_data = photon_nonconservation_data
+
+        for name, data in lightcones.items():
+            setattr(self, name, data)
+
+        # Hold a reference to the global/lightcones in a dict form for easy reference.
+        self.global_quantities = global_quantities
+        self.lightcones = lightcones
+
+    @property
+    def global_xHI(self):
+        """Global neutral fraction function."""
+        warnings.warn(
+            "global_xHI is deprecated. From now on, use global_xH. Will be removed in v3.1"
+        )
+        return self.global_xH
+
+    @property
+    def cell_size(self):
+        """Cell size [Mpc] of the lightcone voxels."""
+        return self.user_params.BOX_LEN / self.user_params.HII_DIM
+
+    @property
+    def lightcone_dimensions(self):
+        """Lightcone size over each dimension -- tuple of (x,y,z) in Mpc."""
+        return (
+            self.user_params.BOX_LEN,
+            self.user_params.BOX_LEN,
+            self.n_slices * self.cell_size,
+        )
+
+    @property
+    def shape(self):
+        """Shape of the lightcone as a 3-tuple."""
+        return self.brightness_temp.shape
+
+    @property
+    def n_slices(self):
+        """Number of redshift slices in the lightcone."""
+        return self.shape[-1]
+
+    @property
+    def lightcone_coords(self):
+        """Co-ordinates [Mpc] of each cell along the redshift axis."""
+        return np.linspace(0, self.lightcone_dimensions[-1], self.n_slices)
+
+    @property
+    def lightcone_distances(self):
+        """Comoving distance to each cell along the redshift axis, from z=0."""
+        return (
+            self.cosmo_params.cosmo.comoving_distance(self.redshift).value
+            + self.lightcone_coords
+        )
+
+    @property
+    def lightcone_redshifts(self):
+        """Redshift of each cell along the redshift axis."""
+        return np.array(
+            [
+                z_at_value(self.cosmo_params.cosmo.comoving_distance, d * units.Mpc)
+                for d in self.lightcone_distances
+            ]
+        )
+
+    def get_unique_filename(self):
+        """Generate a unique hash filename for this instance."""
+        prefix = "LightCone_z{:.4}_".format(self.redshift)
+        rep = ""
+        for inp in [
+            "user_params",
+            "cosmo_params",
+            "astro_params",
+            "flag_options",
+            "global_params",
+        ]:
+            rep += repr(getattr(self, inp))
+
+        rep += (
+            str(np.round(self.node_redshifts, 3))
+            + str(self.global_quantities.keys())
+            + str(self.lightcones.keys())
+        )
+        return prefix + md5(rep.encode()).hexdigest() + ".h5"
+
+    def _write(self, direc=None, fname=None):
+        """
+        Write the lightcone to file in standard HDF5 format.
+
+        This method is primarily meant for the automatic caching. Its default
+        filename is a hash generated based on the input data, and the directory is
+        the configured caching directory.
+
+        Parameters
+        ----------
+        direc : str, optional
+            The directory into which to write the file. Default is the configuration
+            directory.
+        fname : str, optional
+            The filename to write, default a unique name produced by the inputs.
+
+        Returns
+        -------
+        fname : str
+            The absolute path to which the file was written.
+        """
+        direc = os.path.expanduser(direc or config["direc"])
+
+        if fname is None:
+            fname = self.get_unique_filename()
+
+        if not os.path.isabs(fname):
+            fname = os.path.join(direc, fname)
+
+        with h5py.File(fname, "w") as f:
+            # Save input parameters as attributes
+            for k in [
+                "user_params",
+                "cosmo_params",
+                "flag_options",
+                "astro_params",
+                "global_params",
+            ]:
+
+                q = getattr(self, k)
+                kfile = "_globals" if k == "global_params" else k
+                grp = f.create_group(kfile)
+
+                try:
+                    dct = q.self
+                except AttributeError:
+                    dct = q
+
+                for kk, v in dct.items():
+                    if v is None:
+                        continue
+                    try:
+                        grp.attrs[kk] = v
+                    except TypeError:
+                        # external_table_path is a cdata object and can't be written.
+                        pass
+
+            # Save the boxes to the file
+            boxes = f.create_group("lightcones")
+
+            # Go through all fields in this struct, and save
+            for k, val in self.lightcones.items():
+                boxes[k] = val
+
+            global_q = f.create_group("global_quantities")
+            for k, v in self.global_quantities.items():
+                global_q[k] = v
+
+            if self.photon_nonconservation_data is not None:
+                photon_data = f.create_group("photon_nonconservation_data")
+                for k, val in self.photon_nonconservation_data.items():
+                    photon_data[k] = val
+            f.attrs["redshift"] = self.redshift
+            f["node_redshifts"] = self.node_redshifts
+            f.attrs["version"] = __version__
+
+        return fname
+
+    def save(self, fname=None, direc="."):
+        """Save the lightcone to disk.
+
+        This function has defaults that make it easy to save a unique lightcone to
+        the current directory.
+
+        Parameters
+        ----------
+        fname : str, optional
+            The filename to write, default a unique name produced by the inputs.
+        direc : str, optional
+            The directory into which to write the file. Default is the current directory.
+
+        Returns
+        -------
+        str :
+            The filename to which the box was written.
+        """
+        return self._write(direc=direc, fname=fname)
+
+    @classmethod
+    def read(cls, fname, direc=None):
+        """Read a lightcone file from disk, creating a LightCone object.
+
+        Parameters
+        ----------
+        fname : str
+            The filename path. Can be absolute or relative.
+        direc : str
+            If fname, is relative, the directory in which to find the file. By default,
+            both the current directory and default cache and the  will be searched, in
+            that order.
+
+        Returns
+        -------
+        LightCone :
+            A :class:`LightCone` instance created from the file's data.
+        """
+        if not os.path.isabs(fname):
+            fname = os.path.join(direc, fname)
+
+        if not os.path.exists(fname):
+            raise FileExistsError("The file {} does not exist!".format(fname))
+
+        kwargs = {}
+        with h5py.File(fname, "r") as fl:
+            for (k, kls) in [
+                ("user_params", UserParams),
+                ("cosmo_params", CosmoParams),
+                ("flag_options", FlagOptions),
+                ("astro_params", AstroParams),
+            ]:
+                grp = fl[k]
+                kwargs[k] = kls(dict(grp.attrs))
+
+            glbls = dict(fl["_globals"].attrs)
+
+            # Save the boxes to the file
+            boxes = fl["lightcones"]
+
+            kwargs["lightcones"] = {k: boxes[k][...] for k in boxes.keys()}
+
+            glb = fl["global_quantities"]
+            kwargs["global_quantities"] = {k: glb[k][...] for k in glb.keys()}
+
+            if "photon_nonconservation_data" in fl.keys():
+                d = fl["photon_nonconservation_data"]
+                kwargs["photon_nonconservation_data"] = {k: d[k][...] for k in d.keys()}
+
+            kwargs["redshift"] = fl.attrs["redshift"]
+            kwargs["node_redshifts"] = fl["node_redshifts"][...]
+
+        with global_params.use(**glbls):
+            out = cls(**kwargs)
+
+        return out
+
+    def __eq__(self, other):
+        """Determine if this is equal to another object."""
+        return (
+            isinstance(other, self.__class__)
+            and other.redshift == self.redshift
+            and np.all(np.isclose(other.node_redshifts, self.node_redshifts, atol=1e-3))
+            and self.user_params == other.user_params
+            and self.cosmo_params == other.cosmo_params
+            and self.flag_options == other.flag_options
+            and self.astro_params == other.astro_params
+            and self.global_quantities.keys() == other.global_quantities.keys()
+            and self.lightcones.keys() == other.lightcones.keys()
+        )
