+++ src/py21cmfast/plotting.py
@@ -7,8 +7,8 @@ import numpy as np
 from matplotlib import colors
 
 from . import outputs
-from .wrapper import Coeval
-from .wrapper import LightCone
+from .outputs import Coeval
+from .outputs import LightCone
 
 eor_colour = colors.LinearSegmentedColormap.from_list(
     "EoR",
@@ -116,7 +116,7 @@ def coeval_sliceplot(
     Parameters
     ----------
     struct : :class:`~outputs._OutputStruct` or :class:`~wrapper.Coeval` instance
-        The output of a function such as `ionize_box` (a class containing several quantities), or
+        The output of a function such as `ionize_box` (a class containing several lightcones), or
         `run_coeval`.
     kind : str
         The quantity within the structure to be shown.
